+++ packages/core/src/odata/common/request-builder/action-function-import-request-builder-base.ts
@@ -7,12 +7,7 @@ import {
 } from '../../../scp-cf/destination-service-types';
 import { MethodRequestBuilderBase } from '../../common/request-builder/request-builder-base';
 import { ODataFunctionImportRequestConfig } from '../../common/request/odata-function-import-request-config';
-import {
-  FunctionImportParameters,
-  ODataUri,
-  RequestMethodType
-} from '../../common';
-import { ActionImportParameters } from '../../v4/request/action-import-parameter';
+import { ODataActionImportRequestConfig } from '../../v4/request';
 
 /**
  * Create OData request to execute a action or function import.
@@ -23,36 +18,21 @@ export abstract class ActionFunctionImportRequestBuilderBase<
   ParametersT,
   ReturnT
 > extends MethodRequestBuilderBase<
-  ODataFunctionImportRequestConfig<ParametersT>
+  | ODataFunctionImportRequestConfig<ParametersT>
+  | ODataActionImportRequestConfig<ParametersT>
 > {
   /**
    * Base class for function  and actions imports
-   * @param method - HTTP method to be used for the request
-   * @param defaultServicePath - Default path for the service the function belongs to
-   * @param actionOrFunctionImportName - The name of the function import.
    * @param responseTransformer - Transformation function for the response
-   * @param parameters - Parameters to be set in the function
-   * @param oDataUri - Contains the v2/v4 specific URI conversions
+   * @param requestConfig - Request config for a action or funciton import
    */
   protected constructor(
-    method: RequestMethodType,
-    defaultServicePath: string,
-    actionOrFunctionImportName: string,
     readonly responseTransformer: (data: any) => ReturnT,
-    parameters:
-      | FunctionImportParameters<ParametersT>
-      | ActionImportParameters<ParametersT>,
-    oDataUri: ODataUri
+    requestConfig:
+      | ODataFunctionImportRequestConfig<ParametersT>
+      | ODataActionImportRequestConfig<ParametersT>
   ) {
-    super(
-      new ODataFunctionImportRequestConfig(
-        method,
-        defaultServicePath,
-        actionOrFunctionImportName,
-        parameters,
-        oDataUri
-      )
-    );
+    super(requestConfig);
   }
 
   /**