+++ packages/core/src/odata/v2/request-builder/function-import-request-builder.ts
@@ -3,6 +3,7 @@
 import { FunctionImportParameters, RequestMethodType } from '../../common';
 import { oDataUri } from '../uri-conversion';
 import { ActionFunctionImportRequestBuilderBase } from '../../common/request-builder/action-function-import-request-builder-base';
+import { ODataFunctionImportRequestConfig } from '../../common/request/odata-function-import-request-config';
 
 /**
  * Create OData request to execute a function import.
@@ -30,12 +31,14 @@ export class FunctionImportRequestBuilder<
     parameters: FunctionImportParameters<ParametersT>
   ) {
     super(
-      method,
-      defaultServicePath,
-      functionImportName,
       responseTransformer,
-      parameters,
-      oDataUri
+      new ODataFunctionImportRequestConfig(
+        method,
+        defaultServicePath,
+        functionImportName,
+        parameters,
+        oDataUri
+      )
     );
   }
 }