+++ packages/generator/src/edmx-parser/common/edmx-types.ts
@@ -28,7 +28,7 @@ export interface EdmxDocumented {
 
 export interface EdmxParameter extends EdmxDocumented, EdmxNamed {
   Type: string;
-  Nullable: string;
+  Nullable?: string;
 }
 
 export interface EdmxEntitySetBase extends EdmxNamed {