+++ packages/generator/src/edmx-to-vdm/common/function-import.ts
@@ -2,64 +2,38 @@
 import { toTypeNameFormat } from '@sap-cloud-sdk/core';
 import { createLogger } from '@sap-cloud-sdk/util';
 import { ServiceNameFormatter } from '../../service-name-formatter';
-import {
-  VdmComplexType,
-  VdmEntity,
-  VdmFunctionImport,
-  VdmFunctionImportReturnType,
-  VdmFunctionImportReturnTypeCategory
-} from '../../vdm-types';
-import { edmToTsType, isNullableParameter } from '../../generator-utils';
+import { VdmFunctionImportBase } from '../../vdm-types';
 import { SwaggerPath } from '../../swagger-parser/swagger-types';
-import {
-  functionImportDescription,
-  parameterDescription
-} from '../description-util';
-import { EdmxNamed, EdmxParameter } from '../../edmx-parser/common';
-import {
-  filterUnknownEdmTypes,
-  isCollectionType,
-  parseType,
-  parseTypeName,
-  propertyJsType
-} from '../edmx-to-vdm-util';
+import { functionImportDescription } from '../description-util';
+import { EdmxParameter } from '../../edmx-parser/common';
+import { EdmxFunctionImport as EdmxFunctionImportV2 } from '../../edmx-parser/v2';
+import { EdmxFunctionImport as EdmxFunctionImportV4 } from '../../edmx-parser/v4';
+import { getFunctionImportParameters } from './action-function-parameters';
 
 const logger = createLogger({
   package: 'generator',
   messageContext: 'function-import'
 });
 
-export function transformFunctionImportBase<T extends EdmxNamed>(
-  edmxFunctionImport: T,
+export function transformFunctionImportBase(
+  edmxFunctionImport: EdmxFunctionImportV2 | EdmxFunctionImportV4,
   edmxParameters: EdmxParameter[],
   swaggerDefinition: SwaggerPath | undefined,
   formatter: ServiceNameFormatter
-): Omit<VdmFunctionImport, 'httpMethod' | 'returnTypeEdmx' | 'returnType'> {
-  const functionName = formatter.originalToFunctionImportName(
-    edmxFunctionImport.Name
-  );
+): VdmFunctionImportBase {
+  const name = formatter.originalToFunctionImportName(edmxFunctionImport.Name);
   const functionImport = {
     originalName: edmxFunctionImport.Name,
-    functionName,
-    parametersTypeName: toTypeNameFormat(`${functionName}Parameters`)
+    name,
+    parametersTypeName: toTypeNameFormat(`${name}Parameters`)
   };
 
-  const parameters = edmxParameters.filter(filterUnknownEdmTypes).map(p => {
-    const swaggerParameter = swaggerDefinition
-      ? swaggerDefinition.parameters.find(param => param.name === p.Name)
-      : undefined;
-    return {
-      originalName: p.Name,
-      parameterName: formatter.originalToParameterName(
-        edmxFunctionImport.Name,
-        p.Name
-      ),
-      edmType: parseType(p.Type),
-      jsType: edmToTsType(p.Type)!,
-      nullable: isNullableParameter(p),
-      description: parameterDescription(p, swaggerParameter)
-    };
-  });
+  const parameters = getFunctionImportParameters(
+    edmxFunctionImport,
+    edmxParameters,
+    swaggerDefinition,
+    formatter
+  );
 
   return {
     ...functionImport,
@@ -68,107 +42,5 @@ export function transformFunctionImportBase<T extends EdmxNamed>(
       swaggerDefinition,
       functionImport.originalName
     )
-  };
-}
-
-export function parseReturnTypes(
-  returnType: string,
-  entities: VdmEntity[],
-  complexTypes: Omit<VdmComplexType, 'factoryName'>[]
-): VdmFunctionImportReturnType {
-  if (!returnType) {
-    return getVoidReturnType();
-  }
-
-  const isCollection = isCollectionType(returnType);
-  const edmType = findEdmType(returnType);
-  if (edmType) {
-    return getEdmReturnType(isCollection, edmType);
-  }
-
-  const entity = findEntityType(returnType, entities);
-  if (entity) {
-    return getEntityReturnType(isCollection, entity);
-  }
-
-  const complexType = findComplexType(returnType, complexTypes);
-  if (complexType) {
-    return getComplexReturnType(isCollection, complexType);
-  }
-
-  throw Error(`Unable to find a return type for name ${returnType}.`);
-}
-
-function findEdmType(returnType: string): string | undefined {
-  returnType = parseTypeName(returnType);
-  if (returnType.startsWith('Edm.')) {
-    return returnType;
-  }
-}
-
-function findEntityType(
-  returnType: string,
-  entites: VdmEntity[]
-): VdmEntity | undefined {
-  returnType = parseTypeName(returnType);
-  const parsedReturnType = returnType.split('.').slice(-1)[0];
-  return entites.find(e => e.entityTypeName === parsedReturnType);
-}
-
-function findComplexType(
-  returnType: string,
-  complexTypes: Omit<VdmComplexType, 'factoryName'>[]
-): Omit<VdmComplexType, 'factoryName'> | undefined {
-  returnType = parseTypeName(returnType);
-  const parsedReturnType = returnType.split('.').slice(-1)[0];
-  return complexTypes.find(c => c.originalName === parsedReturnType);
-}
-
-function getVoidReturnType(): VdmFunctionImportReturnType {
-  return {
-    returnTypeCategory: VdmFunctionImportReturnTypeCategory.VOID,
-    returnType: 'undefined',
-    builderFunction: '(val) => undefined',
-    isMulti: false,
-    isCollection: false
-  };
-}
-
-function getEdmReturnType(
-  isCollection: boolean,
-  edmType: string
-): VdmFunctionImportReturnType {
-  return {
-    returnTypeCategory: VdmFunctionImportReturnTypeCategory.EDM_TYPE,
-    returnType: propertyJsType(edmType)!,
-    builderFunction: `(val) => edmToTs(val, '${edmType}')`,
-    isMulti: isCollection,
-    isCollection
-  };
-}
-
-function getEntityReturnType(
-  isCollection: boolean,
-  entity: VdmEntity
-): VdmFunctionImportReturnType {
-  return {
-    returnTypeCategory: VdmFunctionImportReturnTypeCategory.ENTITY,
-    returnType: entity.className,
-    builderFunction: entity.className,
-    isMulti: isCollection,
-    isCollection
-  };
-}
-
-function getComplexReturnType(
-  isCollection: boolean,
-  complexType: Omit<VdmComplexType, 'factoryName'>
-): VdmFunctionImportReturnType {
-  return {
-    returnTypeCategory: VdmFunctionImportReturnTypeCategory.COMPLEX_TYPE,
-    returnType: complexType.typeName,
-    builderFunction: `(data) => deserializeComplexType(data, ${complexType.typeName})`,
-    isMulti: isCollection,
-    isCollection
   };
 }