+++ packages/generator/test/function-import/export-statement.spec.ts
@@ -1,10 +1,10 @@
 /* Copyright (c) 2020 SAP SE or an SAP affiliate company. All rights reserved. */
-import { exportStatement } from '../../src/function-import';
 import { orderBreakfast } from '../test-util/data-model';
+import { exportStatement } from '../../src/action-function-import/export-statement';
 
 describe('export-statement', () => {
   it('exportStatement', () => {
-    const actual = exportStatement([orderBreakfast]);
+    const actual = exportStatement([orderBreakfast], 'functionImports');
 
     expect(actual.declarationKind).toEqual('const');
     expect(actual.isExported).toBeTruthy();