+++ packages/generator/test/function-import/response-transformer-function.spec.ts
@@ -1,20 +1,21 @@
 /* Copyright (c) 2020 SAP SE or an SAP affiliate company. All rights reserved. */
-import { responseTransformerFunctionName } from '../../src/function-import';
-import { VdmFunctionImportReturnTypeCategory } from '../../src/vdm-types';
+
+import { VdmReturnTypeCategory } from '../../src/vdm-types';
 import { orderBreakfast } from '../test-util/data-model';
+import { responseTransformerFunctionName } from '../../src/action-function-import/response-transformer-function';
 
 const returnTypeEntity = {
   builderFunction: '',
   returnType: 'Entity',
   isCollection: false,
-  returnTypeCategory: VdmFunctionImportReturnTypeCategory.ENTITY
+  returnTypeCategory: VdmReturnTypeCategory.ENTITY
 };
 
 const returnTypeComplexType = {
   builderFunction: '',
   returnType: 'ComplexType',
   isCollection: true,
-  returnTypeCategory: VdmFunctionImportReturnTypeCategory.COMPLEX_TYPE
+  returnTypeCategory: VdmReturnTypeCategory.COMPLEX_TYPE
 };
 
 describe('response-transformer-function', () => {