+++ packages/generator/test/generator.spec.ts
@@ -67,7 +67,7 @@ describe('generator', () => {
     });
 
     it('generates expected number of files', () => {
-      expect(files.length).toBe(29);
+      expect(files.length).toBe(30);
     });
 
     it('generates TestEntity.ts file', () => {
@@ -98,6 +98,20 @@ describe('generator', () => {
       );
       expect(functionImportNames).toEqual(
         expect.not.arrayContaining(['testFunctionImportNoReturnType'])
+      );
+    });
+
+    it('generates action-imports.ts file', () => {
+      const actionImports = getActionImportDeclarations(files);
+      expect(actionImports.length).toBe(2);
+      const actionImportNames = actionImports.map(action => action.getName());
+      expect(actionImportNames).toEqual(
+        expect.arrayContaining(['testActionImportNoParameterNoReturnType'])
+      );
+      expect(actionImportNames).toEqual(
+        expect.arrayContaining([
+          'testActionImportMultipleParameterComplexReturnType'
+        ])
       );
     });
   });
@@ -141,4 +155,14 @@ function getFunctionImportDeclarations(
   );
 
   return functionImportsFile!.getFunctions();
+}
+
+function getActionImportDeclarations(
+  files: SourceFile[]
+): FunctionDeclaration[] {
+  const actionImportFile = files.find(
+    file => file.getBaseName() === 'action-imports.ts'
+  );
+
+  return actionImportFile!.getFunctions();
 }