+++ CHANGELOG.md
@@ -20,6 +20,7 @@
 ## New Functionality
 
 - Support collections in complex types.
+- Support unbound actions in OData V4. Note that all OData V4 functionality is still experimental.
 
 ## Improvements